refactor: mejora de tipado
