+++ app/page.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { BoxModel } from "@/components/BoxModel";
 import { Navbar } from "@/components/Navbar";
 import { SearchModels } from "@/components/SearchModels";
-import { ModelData } from "@/utils/types";
+import { ModelData } from "@/types/types";
 
 export default function Home() {
   const [dataModels, setDataModels] = useState<ModelData[]>([]);
