+++ types/types.ts
@@ -1,0 +1,27 @@
+export interface ModelData {
+  nombreCompleto: string;
+  edad: string;
+  descripcion: string;
+  etnia: string;
+  zona: string;
+}
+
+export interface DataModels {
+  activo: boolean;
+  ciudad: string;
+  createdAt: string;
+  descripcion: string;
+  duracionesAdicionales: string;
+  edad: number;
+  email: string;
+  etnia: string;
+  fechaNacimiento: string;
+  idiomas: string[];
+  multimedias: string[];
+  nombreCompleto: string;
+  nombreUsuario: string;
+  numeroContacto: string;
+  precioHora: number;
+  rol: string;
+  zona: string;
+}
