--- utils/types.ts
@@ -1,7 +1,0 @@
-export interface ModelData {
-  nombreCompleto: string;
-  edad: string;
-  descripcion: string;
-  etnia: string;
-  zona: string;
-}
